docs: turn on ssg and ssg-md in rslint playground
